requirements.txt
streamlit 
langchain
langchain-google-genai 
python-dotenv
langchain_community
langchain_openai
